main.py changed phrasing of some comments
input_generator.py implemented get_coordinates_from_xyz method which allowed generate_input_file method to be modified such that the input files that are generated are filled with the xyz coordinates block rather than the filename. Allows for easier basis set specification when editing the raw output manually.

adjusted the configuration readers get_molecular_configurations method so that it would grab sections that dont start with "calculation" as opposed to grabbing ones that start with 'molecule' because I wanted the output to be named apprpropiately. output also isnt prefixed with output_ anymore.

create_configuration_file.py had three changes, a typo in the user input section, I changed the section name generation and prefixed the structure_file key entries with "xyzdata\\" to be sure of absolute path specification